Break longer test names for better rendering
For longer test names, failure cell is not
immediately visible in the html report
without scrolling to right as the
container table expands as per test name length.
Break longer test names to limit the width of
first cell so that failure cell is immediately
visible after opening the html report without
the need for scrolling.